+++ schedulerx2-20190430/ChangeLog.md
@@ -1,3 +1,9 @@
+2025-10-27 Version: 2.1.2
+- Update API CreateAppGroup: add request parameters NotificationPolicyName.
+- Update API GetAppGroup: add response parameters Body.Data.NotificationPolicyName.
+- Update API UpdateAppGroup: add request parameters NotificationPolicyName.
+
+
 2025-06-10 Version: 2.2.0
 - Support API ReadSchedulerxDesignateInfo.